#56777 게시판 댓글 권한 처리
•게시글 수정 삭제와 마찬가지로 게시판 댓글에 대해서도 댓글 수정 삭제에 대한 권한 추가

http://iscr01.navercorp.com/issues/56777

+++ src/main/java/com/naver/jihyunboard/reply/controller/ReplyController.java
@@ -58,14 +58,25 @@ public class ReplyController {
 
 	@ResponseBody
 	@RequestMapping(value = "/update", method = RequestMethod.POST)
-	public void update(Reply reply, Model model) {
-		replyService.updateReply(reply);
+	public String update(Reply reply, Authentication auth, Model model) {
+		if (boardService.authUserId(auth).equals(replyService.replyerId(reply.getReplyNum()))) {
+			replyService.updateReply(reply);
+			return "ok";
+		} else {
+			return "permission denied";
+		}
+
 	}
 
 	@ResponseBody
-	@RequestMapping(value = "/delete", method = RequestMethod.GET)
-	public void delete(Reply reply, Model model) {
-		replyService.deleteReply(reply);
+	@RequestMapping(value = "/delete", method = RequestMethod.GET, produces = "text/plain")
+	public String delete(Reply reply, Authentication auth, Model model) {
+		if (boardService.authUserId(auth).equals(replyService.replyerId(reply.getReplyNum()))) {
+			replyService.deleteReply(reply);
+			return "ok";
+		} else {
+			return "permission denied";
+		}
 	}
 
 }
